Move `CustomDns` into `ServiceEndpoint`
