+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/MullvadVpnService.kt
@@ -235,10 +235,6 @@ class MullvadVpnService : TalpidVpnService() {
     }
 
     private suspend fun setUpInstance(daemon: MullvadDaemon, settings: Settings) {
-        val customDns = CustomDns(endpoint)
-
-        endpoint.customDns = customDns
-
         handlePendingAction(settings)
 
         if (state == State.Running) {
